requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
cloudinary==1.39.0
Django==4.2.10
django-autoslug==1.9.9
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==24.0
phonenumberslite==8.13.30
pillow==10.2.0
psycopg2==2.9.9
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
ruff==0.5.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
